Fix String bug. Add some model definitions

+++ include/String.cpp
@@ -417,7 +417,7 @@ String String::toString(double num, unsigned short precision) {
 
 	res += numToString(num);
 
-	num -= (unsigned int)num;
+	num -= (unsigned long long)num;
 
 	if (num == 0 || precision == 0) {
 		return res;
